requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.3.0
bitsandbytes==0.47.0
certifi==2025.8.3
charset-normalizer==3.4.3
comm==0.2.3
contourpy==1.3.3
cut-cross-entropy==25.1.1
cycler==0.12.1
datasets==3.6.0
debugpy==1.8.16
decorator==5.2.1
dill==0.3.8
distro==1.9.0
docstring_parser==0.17.0
executing==2.2.1
filelock==3.19.1
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.3.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.9
hf_transfer==0.1.9
hickle==5.0.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
inquirerpy==0.3.4
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.4
jiter==0.11.0
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
lightning-utilities==0.15.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgspec==0.19.0
multidict==6.6.4
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.3
numpy==2.3.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
openai==2.3.0
packaging==25.0
pandas==2.3.2
parso==0.8.5
peft==0.17.1
pexpect==4.9.0
pfzy==0.3.4
pillow==11.0.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
sentencepiece==0.2.1
setuptools==78.1.1
shtab==1.7.2
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.13.3
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.0
torch==2.7.1+cu118
torchao==0.13.0
torchaudio==2.7.1+cu118
torcheval==0.0.7
torchmetrics==1.8.2
torchvision==0.22.1+cu118
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.1
triton==3.3.1
trl==0.22.2
typeguard==4.4.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tyro==0.9.31
tzdata==2025.2
unsloth @ git+https://github.com/unslothai/unsloth.git@551bc65235aaf4ce1d00b48fed79eb48d759507b
unsloth_zoo==2025.9.2
urllib3==2.5.0
wcwidth==0.2.13
wheel==0.45.1
xxhash==3.5.0
yarl==1.20.1
